Refactoring to allow canvas placement and setup at arbitrary nodes.
Added container arg to loadData to allow caller to specify the parent element
for the canvas. If the argument is not supplied then it defaults to the old
behavior.

Refactored setup such that setup(container) should be called if you want
to parse the taxonomy name and the node ID from the URL. If the caller
wants to supply these as arguments, then:
    setupAtNode(nodeid, domsource, container)
should be used.